Refactor Client to AbstractClient and simplify implementation
Replaces the Client class with a new AbstractClient class, removing SSL context and certificate management logic. The new abstract class focuses on HTTP request handling and logging, streamlining the client interface for HTTP operations.
