--- src/main/java/nl/radiantrealm/library/http/client/CertificateLevel.java
@@ -1,7 +1,0 @@
-package nl.radiantrealm.library.http.client;
-
-public enum CertificateLevel {
-    PUBLIC_CA,
-    INTERNAL_CA,
-    INSECURE_CA;
-}
